+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- `update_logger` function.
 - Customized wrapper for ranger package.
 - Extended plot function for gen_pseudo_pop object (plot.R)
 - Extended plot function for estimate_erf object (plot.R)